ci(openclaw): split the host-contract pin — pi-ai on every push, openclaw in its own job
The dual-review consensus: `@mariozechner/pi-ai` STAYS an exact devDependency;
`openclaw` stops being one and its contract moves to a dedicated CI job.

The two hosts are not comparable dependencies. pi-ai is small and the
assertions that read it are cheap, so they keep running in-line on every push.
openclaw is a full agent CLI: as a devDependency it pulled ~309 packages into
every single `npm ci` in the workflow — four jobs' worth — to serve four
type-only imports in two test files. The contract only ever needs to be proven
once per run. So it is, in `openclaw-contract`.

What actually keeps openclaw out of the install is worth recording, because it
is not the devDependency removal on its own. openclaw remains an OPTIONAL PEER
of `packages/openclaw` — it must, that is the plugin's real relationship to its
host — and npm resolves optional peers into the lockfile whether or not anyone
asked. The entry is therefore still there, now flagged `peer`/`optional`
instead of `dev`. What changed is that **`npm ci` does not install optional
peers**: verified against a clean tree, 260 packages installed and no
`node_modules/openclaw`, versus 569 resolved. Every ordinary CI job now runs
without openclaw on disk.

Which forced the second half of the design. `tsc` cannot skip a file whose
import does not resolve, so a conditional include was never available; the
openclaw assertions had to move to their own project. The split, and the
guarantee that nothing fell through it:

  - tests/contract.test-d.ts (31 assertions, pi-ai + our own mirror)
    -> tsconfig.type-tests.json, now wired into `npm run typecheck`, so it
       gates EVERY push. It gated nothing before this commit — that config was
       referenced by no CI step at all, a gap AGENTS.md recorded and this fixes.
  - tests/contract-openclaw.test-d.ts (32 assertions) + contract.test.ts
    -> tsconfig.contract-openclaw.json, compiled only by the new job.

31 + 32 = the 63 assertions the single file carried; the exported-name sets are
identical before and after. Every assertion still runs somewhere in CI.

contract.test.ts now has two modes and no third. Under
`USERTRUST_OPENCLAW_CONTRACT=1` — the new job — an absent or version-MISMATCHED
openclaw throws before any test runs, because a job whose whole purpose is
proving a pin must never pass by proving nothing. Everywhere else the suite
skips with a notice naming that job as the proving ground. The notice goes to
`process.stderr` rather than `console.warn`: vitest attributes intercepted
console output to a running task and its default reporter drops what a
skipped-file module load produced, which made the first version of this silent
instead of loud.

The pin lives in exactly ONE file, `packages/openclaw/openclaw-contract.env`,
sourced by the job and re-read by the test for the mismatch check. Three doc
comments and a README paragraph that restated the version — one of which now
claimed something false — point at it instead.

One deviation from the specified job, with evidence. `--package-lock=false`
makes arborist rebuild the ideal tree from scratch and npm 10.9.2 then crashes
on this repo's peer sets (`TypeError: Cannot read properties of null (reading
'edgesOut')` in `#loadPeerSet`), reproducibly, on a tree straight out of
`npm ci`. `--no-save` alone does the job: verified byte-identical package.json,
packages/*/package.json and package-lock.json across the install, and the job
now asserts that with `git diff --exit-code` rather than trusting it.

Separately, `tb-integration` hard-coded one suite. It now runs all three
real-cluster suites, naming each file explicitly — `envelope-integration.test.ts`
lives outside any `integration/` folder, so a directory glob would have gone on
silently missing it and `budget-envelope.tb.test.ts`. Both are new on this
branch and had never run in CI.

Verified locally: the full `openclaw-contract` job end-to-end from a clean
`npm ci`; the full suite and the CI coverage gate with openclaw ABSENT (the
real CI state) and present; all three TB suites against a live cluster.
