+++ packages/openclaw/openclaw-contract.env
@@ -1,0 +1,23 @@
+# The exact openclaw release this plugin's host contract is pinned to.
+#
+# THE SINGLE SOURCE OF THIS PIN. It is deliberately NOT a devDependency of
+# packages/openclaw: openclaw is a full agent CLI, and pulling its whole tree
+# into every `npm ci` — for four type-only imports in two test files — bought a
+# large install and a large lockfile surface for a contract that only ever needs
+# to be proven once per run. (`@mariozechner/pi-ai` STAYS an exact devDependency:
+# it is small, and the assertions that read it run on every push.)
+#
+# Who reads this file:
+#   - .github/workflows/ci.yml, job `openclaw-contract` — sources it, installs
+#     exactly this version out-of-tree (`--no-save --package-lock=false`), and
+#     runs the openclaw half of the contract with USERTRUST_OPENCLAW_CONTRACT=1.
+#     That job is the proving ground; nothing else in CI sees openclaw at all.
+#   - packages/openclaw/tests/contract.test.ts — reads it to hard-fail on a
+#     version MISMATCH, so an install that silently resolved something else
+#     cannot pass as a proof of this pin.
+#
+# Rationale for every field name the contract asserts on: contract-notes §2.
+# Bumping the pin: change the value here, then run the `openclaw-contract` job's
+# commands locally and fix whatever the compile reports. Nothing else references
+# the version number.
+OPENCLAW_CONTRACT_VERSION=2026.7.1-2
